+++ src/index.ts
@@ -1,5 +1,7 @@
 export * from "./button/index.js";
 export * from "./core/index.js";
+export * from "./count/index.js";
 export * from "./effects/index.js";
 export * from "./icon/index.js";
 export * from "./progress/index.js";
+export * from "./text/index.js";